Update motor controller

+++ include/platform.h
@@ -11,9 +11,9 @@
 #define ENCODER_UPDATE_RATE 100
 
 typedef struct{
-    int8_t x;
-    int8_t y;
-    int8_t t;
+    double x;
+    double y;
+    double t;
 } platform_velocity_t;
 
 void initialize_mecanum_platform(uint8_t isReversed0, uint8_t isReversed1, uint8_t isReversed2, uint8_t isReversed3);
